+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 0.1.2
 
 Improvements to parsing logic for responses:
+
 * If response struct members are in snake-case - Go struct should have member in camel-case
 * It is now possible to use type aliases when decoding a response (#1)
+* It is now possible to use pointers to fields, without getting an error (#2)
 
 ## 0.1.1